Started decompiling `ef_util` Almost all functions are decompiled (but not added to the CPP yet, they're in decomp.me) except for: - void MtxGetRotation(const MTX34& param_1, VEC3* param_2): https://decomp.me/scratch/lSNNS - void MtxGetScale(const MTX34& param_1, VEC3* param_2): https://decomp.me/scratch/byk0f
Because MtxGetScale() has not been decompiled yet, I can't add the rest of the matched functions yet because otherwise I will get undefined errors with some of the labels on .sdata2, regarding the MtxGetScale() assembly (and if I keep those, then the output won't match due to creating new additional data on .sdata2).

MtxGetScale() has been a bit difficult to decompile for me due to all the inline functions it uses.